# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: SSCMS API
  description: 'Smart Student Club Management System -- Backend Sprint 1 (Milestone
    3) covered Epics 1-3: Role-Based Portal Access & Onboarding, Venue Booking & Asset
    Management, and Event Lifecycle & Approval Workflows. Backend Sprint 2 (Milestone
    4) adds Epics 4-6: Task Management & Volunteer Workflows, QR Attendance & Live
    Execution, and Certification & Budget Analytics.'
  version: 1.1.0
paths:
...
  /notifications:
    get:
      tags:
      - Notifications
      summary: List My Notifications
      operationId: list_my_notifications_notifications_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NotificationOut'
                type: array
                title: Response List My Notifications Notifications Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - Notifications
      summary: Create Notification
      operationId: create_notification_notifications_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /notifications/read-all:
    post:
      tags:
      - Notifications
      summary: Read All Notifications
      operationId: read_all_notifications_notifications_read_all_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
...
  /llm/notification-digest:
    post:
      tags:
      - LLM-Assisted Features
      summary: Notification Digest
      description: 'Cross-cutting platform feature backing every role''s notification
        feed: summarizes a caller''s unread notifications (approvals, rejections,
        deadlines) into a short digest plus up to 3 highlight bullets. Calls Gemini
        (GEMINI_API_KEY) when configured; otherwise falls back to a deterministic
        count-based summary -- `source` in the response tells the caller which path
        was used.'
      operationId: notification_digest_llm_notification_digest_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationDigestRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDigestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /llm/event-proposal-assist:
    post:
      tags:
      - LLM-Assisted Features
      summary: Event Proposal Assistant
      description: 'Supports Story 3.1 (Event Proposal Pipeline): when a Club Head
        is creating a new event proposal, this endpoint takes a minimal input (event
        name, objective, expected participants) and returns a fully drafted proposal
        including: event description, agenda, budget estimate, required inventory,
        volunteer requirements, risk assessment, and a planning timeline. Uses Gemini
        when GEMINI_API_KEY is set; otherwise falls back to a deterministic heuristic.
        The response''s `source` field indicates which path was used.'
      operationId: event_proposal_assist_llm_event_proposal_assist_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventProposalAssistRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventProposalAssistResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
    AgendaItem:
      properties:
        time_slot:
          type: string
          title: Time Slot
          description: e.g. '10:00 - 10:30'
        activity:
          type: string
          title: Activity
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - time_slot
      - activity
      title: AgendaItem
...
    EventProposalAssistRequest:
      properties:
        event_name:
          type: string
          minLength: 1
          title: Event Name
          description: Name / title of the proposed event
        objective:
          type: string
          minLength: 1
          title: Objective
          description: Purpose or goal of the event
        expected_participants:
          type: integer
          exclusiveMinimum: 0.0
          title: Expected Participants
          description: Estimated number of attendees
      type: object
      required:
      - event_name
      - objective
      - expected_participants
      title: EventProposalAssistRequest
      description: Club Head provides the basic idea; AI generates the full proposal
        content.
    EventProposalAssistResponse:
      properties:
        source:
          type: string
          title: Source
          description: '"gemini" if the LLM produced the content, "heuristic" otherwise'
        event_description:
          type: string
          title: Event Description
        agenda:
          items:
            $ref: '#/components/schemas/AgendaItem'
          type: array
          title: Agenda
          default: []
        budget_estimate:
          type: number
          title: Budget Estimate
          description: Estimated total budget in INR
        required_inventory:
          items:
            $ref: '#/components/schemas/InventoryItemSuggestion'
          type: array
          title: Required Inventory
          default: []
        volunteers_required:
          type: integer
          title: Volunteers Required
        risk_assessment:
          type: string
          title: Risk Assessment
        timeline:
          items:
            $ref: '#/components/schemas/TimelinePhase'
          type: array
          title: Timeline
          default: []
      type: object
      required:
      - source
      - event_description
      - budget_estimate
      - volunteers_required
      - risk_assessment
      title: EventProposalAssistResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InventoryItemSuggestion:
      properties:
        item:
          type: string
          title: Item
        quantity:
          type: integer
          title: Quantity
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - item
      - quantity
      title: InventoryItemSuggestion
...
    NotificationCreate:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: NotificationCreate
    NotificationDigestRequest:
      properties:
        notifications:
          items:
            $ref: '#/components/schemas/NotificationIn'
          type: array
          title: Notifications
      type: object
      required:
      - notifications
      title: NotificationDigestRequest
    NotificationDigestResponse:
      properties:
        source:
          type: string
          title: Source
          description: '"gemini" if the LLM produced the digest, "heuristic" otherwise'
        summary:
          type: string
          title: Summary
        highlights:
          items:
            type: string
          type: array
          title: Highlights
          default: []
      type: object
      required:
      - source
      - summary
      title: NotificationDigestResponse
    NotificationIn:
      properties:
        message:
          type: string
          title: Message
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      type: object
      required:
      - message
      title: NotificationIn
    NotificationOut:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        message:
          type: string
          title: Message
        read:
          type: boolean
          title: Read
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - user_id
      - message
      - read
      - created_at
      title: NotificationOut
...
    TimelinePhase:
      properties:
        phase:
          type: string
          title: Phase
        start_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
          - type: 'null'
          title: End Date
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - phase
      title: TimelinePhase
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login